+++ CHANGELOG.md
@@ -5,3 +5,6 @@ All notable changes to this project will be documented in this file.
 ## 21-01-2026 - Petr Valenta
 * Merge of working base of the program
 * Base functions are working
+
+## 21-01-2026 - Petr Valenta
+* Safe multiprocess logging to file